+++ thesimulator/util/cspaces/boost_graph_space.h
@@ -74,7 +74,7 @@ namespace cstuff {
         }
 
         GraphSpace(double velocity, vector <vertex_t> vertex_vec, vector <Edge> edge_vec)
-                : GraphSpace(velocity, vertex_vec, edge_vec, vector < double > {edge_vec.size(), 1}) {}
+                : GraphSpace(velocity, vertex_vec, edge_vec, vector < double > {static_cast<int>(edge_vec.size()), 1}) {}
 
         GraphSpace(double velocity, vector <vertex_t> vertex_vec, vector <Edge> edge_vec, vector<double> weight_vec)
                 : _g{vertex_vec.size()}, velocity{velocity}, vertex2label{get(vertex_name, _g)},
